requirements.txt
pbr>=0.5.21,<0.7.0
Babel>=0.9.6

alembic>=0.4.1
oslo.config>=1.2.0
PasteDeploy>=1.5.0
pecan
python-keystoneclient>=0.4.1
redis
six>=1.4.1
SQLAlchemy>=0.7.8,<0.7.99
sqlalchemy-migrate>=0.8.2
wsme>=0.5b2

eventlet>=0.12.0
iso8601
netaddr
